+++ yiffspace-user/test/test_helper.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+require("yiffspace/user")
+require("active_support/test_case")
+require("minitest/autorun")