+++ yiffspace-config/CHANGELOG.md
@@ -1,3 +1,3 @@
 ## 0.0.1
 
-- Initial release, extracted from the `yiffspace` gem's `YiffSpace::ConfigBuilder`.
+- Initial release, extracted from the `yiffspace` gem's `YiffSpace::Config::Builder`.